--- src/helper-functions/show-super_user-options.js
@@ -1,9 +1,0 @@
-export const showSuperUserOption = (setShowOption, isOptionKey) => {
-  if (isOptionKey) {
-    setShowOption(true);
-  }
-};
-
-export const hideSuperUserOption = (setShowOption) => {
-  setShowOption(false);
-};
